增加Map实现

+++ src/main/java/com/lamfire/pandora/FireCollection.java
@@ -2,7 +2,7 @@ package com.lamfire.pandora;
 
 public interface FireCollection {
 
-    public long size();
+    public int size();
 
     public void clear();
