+++ src/main/java/com/lamfire/pandora/LDBFireMap.java
@@ -117,10 +117,10 @@ class LDBFireMap implements FireMap {
     }
 
     @Override
-    public long size() {
+    public int size() {
         try{
             lock.lock();
-            return meta.getValueAsLong(this.sizeKey);
+            return (int)meta.getValueAsLong(this.sizeKey);
         }finally {
             lock.unlock();
         }
