+++ src/main/java/com/lamfire/pandora/LDBFireRank.java
@@ -361,10 +361,10 @@ class LDBFireRank implements FireRank {
     }
 
     @Override
-    public long size() {
+    public int size() {
         try {
             lock.lock();
-            return meta.getValueAsLong(this.sizeKey);
+            return (int)meta.getValueAsLong(this.sizeKey);
         } finally {
             lock.unlock();
         }
