+++ src/main/java/com/lamfire/pandora/LDBFireSet.java
@@ -171,10 +171,10 @@ class LDBFireSet implements FireSet {
     }
 
     @Override
-    public long size() {
+    public int size() {
         try {
             lock.lock();
-            return meta.getValueAsLong(this.sizeKey);
+            return (int)meta.getValueAsLong(this.sizeKey);
         } finally {
             lock.unlock();
         }
